+++ tests/Factory/Header/FactoriesTest.php
@@ -7,7 +7,7 @@ use Innmind\Http\Factory\{
     Header\Factories,
     Header\DelegationFactory
 };
-use Innmind\TimeContinuum\Earth\Clock;
+use Innmind\TimeContinuum\Clock;
 use Innmind\Immutable\Map;
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
@@ -15,18 +15,18 @@ class FactoriesTest extends TestCase
 {
     public function testAll()
     {
-        $all = Factories::all(new Clock);
+        $all = Factories::all(Clock::live());
 
         $this->assertInstanceOf(Map::class, $all);
         $this->assertCount(26, $all);
-        $this->assertEquals($all, Factories::all(new Clock));
+        $this->assertEquals($all, Factories::all(Clock::live()));
     }
 
     public function testDefault()
     {
-        $factory = Factories::default(new Clock);
+        $factory = Factories::default(Clock::live());
 
         $this->assertInstanceOf(DelegationFactory::class, $factory);
-        $this->assertEquals($factory, Factories::default(new Clock));
+        $this->assertEquals($factory, Factories::default(Clock::live()));
     }
 }