+++ tests/Header/DateValueTest.php
@@ -7,14 +7,16 @@ use Innmind\Http\Header\{
     DateValue,
     Value
 };
-use Innmind\TimeContinuum\Earth\PointInTime\PointInTime;
+use Innmind\TimeContinuum\PointInTime;
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
 class DateValueTest extends TestCase
 {
     public function testInterface()
     {
-        $h = new DateValue(new PointInTime('2016-01-01 12:12:12+0200'));
+        $h = new DateValue(PointInTime::at(
+            new \DateTimeImmutable('2016-01-01 12:12:12+0200'),
+        ));
 
         $this->assertInstanceOf(Value::class, $h);
         $this->assertSame('Fri, 01 Jan 2016 10:12:12 GMT', $h->toString());